update the .pyx and test file

+++ python/cuml/ensemble/randomforestclassifier.pyx
@@ -177,7 +177,8 @@ cdef extern from "randomforest/randomforest.hpp" namespace "ML":
                                     bool,
                                     bool,
                                     int,
-                                    float, CRITERION,
+                                    float,
+                                    CRITERION,
                                     bool,
                                     int) except +
 
@@ -481,7 +482,7 @@ class RandomForestClassifier(Base):
                                      <bool> self.bootstrap_features,
                                      <bool> self.bootstrap,
                                      <int> self.n_estimators,
-                                     <int> self.rows_sample,
+                                     <float> self.rows_sample,
                                      <CRITERION> self.split_criterion,
                                      <bool> self.quantile_per_tree,
                                      <int> self.n_streams)
